requirements.txt
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
matplotlib==3.8.2
Pillow==10.2.0
